backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
praw
transformers
torch
PyYAML
